+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## WIP
 
 - Prevent uploading .phps + .html files in assets manager
+- Require verification for updating user data
 
 ## 2.6.2 (2023-07-31)